Fix a bug related to an overflow correction
When porting from dlmalloc.c this location was altered from the original source
because the original version contained an integer overflow. It turns out,
though, that the correctness of the check actually relied on integer overflow
happening! This commit tweaks this check to account for this and bypass the
statement entirely if integer overflow would have happened.

This should help fix a pathological use case seen where lots of memory was
free'd but it was never reused.