xous: add dlmalloc backend for Xous
Xous is a Rust Tier-3 operating system. This microkernel system contains
no C code, and therefore has no default allocator.

dlmalloc was chosen as an initial implementation because it is simple to
understand and does not require much glue.

The Xous `IncreaseHeap` syscall is analagous to the wasm `memory_grow()`
call, and behaves much in the same way.

Like wasm, `free()` is unimplemented.

A future version of this may use the `MapMemory()` and `UnmapMemory()`
syscalls instead of the Heap API.

Signed-off-by: Sean Cross <sean@xobs.io>