Update to windows-sys v0.59
This is the first new version since v0.52. Since
dlmalloc compiles and passes tests without any
changes, I relaxed the version requirement to
allow both versions. That should contribute a
little to avoiding lockfiles with two windows-sys
versions in them. Such duplication is pretty
harmless in the case of dlmalloc (since it's
rarely if ever built for windows by dependents),
but at the very least it still trips up tools like
cargo-deny that scan the lockfile.