Don't allocate the final addresses on wasm
This is intended at being one fix to rust-lang/rust#144199. It's not
valid in Rust for an allocation to include the final byte of the address
space, so when the final wasm page is allocated it cannot be entirely in
use. This fakes the last 16 bytes as un-allocate-able and this is
sufficient for resolving rust-lang/rust#144199.